refactor: extract functions
